# routers/api/room/roomController.go
package room

import (
	"fmt"
	"github.com/gin-gonic/gin"
	"go_sugared/pkg/util"
	"go_sugared/routers/api"
	"gopkg.in/mgo.v2/bson"
	"time"
)

// @Summary 新增房源信息
// @Tags 房源相关
// @Accept application/json
// @Param name query string true "姓名"
// @Param state query int false "State"
// @Param created_by query int false "CreatedBy"
// @Success 200 {string} json "{"code":200,"data":{},"msg":"ok"}"
// @Router /hh/room/add [post]
func AddRoom(c *gin.Context) {
	fmt.Println("add room")
	roomAddReq := &Room{
		RenterId: make([]string,0),
		TagId: make([]string,0),
	}
	if err := c.BindJSON(&roomAddReq); err != nil {
		api.ApiBaseResponse(c, 400, "request param error...")
	} else {
		roomAddReq.PackageNumber = util.MakePackageNumber()
		roomAddReq.ToPrint()
		roomAddReq.UpdateTime = time.Now()
		roomAddReq.ReleaseTime = time.Now()
		if err := roomAddReq.Insert(); err != nil {
			fmt.Println("insert err: ", err)
			api.ApiBaseResponse(c, 500, "insert room error...")
		} else {
			res := &Room{
				RenterId: make([]string, 0),
				TagId:    make([]string, 0),
			}
			res, _ = FindOneRoomByPackageNumber(roomAddReq.PackageNumber)
			api.ApiDataResponse(c, 200, "insert room seccess...", res)
		}
	}
}



// @Summary 删除房源信息
// @Tags 房源相关
// @Accept application/json
// @Param id query string true "房源id"
// @Param state query int false "State"
// @Param created_by query int false "CreatedBy"
// @Success 200 {string} json "{"code":200,"data":{},"msg":"ok"}"
// @Router /hh/room/delete [post]
// 删除房源信息
func DeleteRoom(c *gin.Context) {
	fmt.Println("delete room")
	roomDeleteReq := &SingleRoomReq{}
	if err := c.BindJSON(&roomDeleteReq); err != nil {
		api.ApiBaseResponse(c, 400, "request param error...")
	} else {
		if err := roomDeleteReq.Delete(); err != nil {
			fmt.Println("delete err: ", err)
			api.ApiBaseResponse(c, 500, "delete room error...")
		} else {
			api.ApiBaseResponse(c, 200, "delete room seccess...")
		}
	}
}

// @Summary 更新房源信息
// @Tags 房源相关
// @Accept application/json
// @Param id query string true "房源id"
// @Param state query int false "State"
// @Param created_by query int false "CreatedBy"
// @Success 200 {string} json "{"code":200,"data":{},"msg":"ok"}"
// @Router /hh/room/update [post]
// 编辑房源信息
func UpdateRoom(c *gin.Context) {
	roomUpdateReq := &Room{}
	if err := c.BindJSON(&roomUpdateReq); err != nil {
		api.ApiBaseResponse(c, 400, "request param error...")
	} else {
		update := MakeRoomUpdate(*roomUpdateReq)
		res, err := UpdateRoomBySelector("hh", "room", update, roomUpdateReq.PackageNumber)
		if err != nil {
			fmt.Println(err)
			api.ApiBaseResponse(c, 500, "update error...")
		} else {
			api.ApiDataResponse(c, 200, "update room seccess...", res)
		}
	}
}

// @Summary 获取某一个房源具体信息
// @Tags 房源相关
// @Accept application/json
// @Param id query string true "房源id"
// @Param state query int false "State"
// @Param created_by query int false "CreatedBy"
// @Success 200 {string} json "{"code":200,"data":{},"msg":"ok"}"
// @Router /hh/room/getDetail [post]
// 获取具体房源信息 通过筛选条件
func GetRoomDetail(c *gin.Context) {
	fmt.Println("get room detail")
	getreq := &SingleRoomReq{}
	if err := c.BindJSON(&getreq); err != nil {
		api.ApiBaseResponse(c, 500, "request param error!!!")
	} else {
		fmt.Println("request: ", getreq)
		req := MakeSelector(*getreq)
		res := []Room{
			{
				RenterId: make([]string,0),
				TagId: make([]string,0),
			},
		}
		err := api.FindAllBySelector("hh", "room", req, bson.M{"_id": 0}, &res)
		if err != nil || len(res)==0{
			api.ApiDataResponse(c, 404, "find one room error or not find...", nil)
		} else {
			api.ApiDataResponse(c, 200, "find one room seccess...", res)
		}
	}
}

// 获取全部房源信息
func GetRoomIndex(c *gin.Context) {
	fmt.Println("get all room list....")
	var res []Room
	err := api.FindAllBySelector("hh", "room", nil, bson.M{"_id": 0}, &res)
	if err != nil {
		api.ApiBaseResponse(c, 500, "find all room error...")
	} else {
		api.ApiDataResponse(c, 200, "index seccess...", res)
	}
}